Fix flaw in http_pipe. Sessions for the server-end improperly ordered.
The amazing part is this mostly worked. Using lists:keyfind works fine
for checking which session is the first one for the client-side.
However, this does not work well at all for checking which session is
the currently active session for the server-side of the pipe(s). Added a
dictionary to keep track of the order in which a proc listen-s to
session Ids.

Changed the mechanics of cancels and resets. These now send a message to
http_res_sock or http_req_sock. These processes must decide on their own
whether they should exit, in case their pipeline cannot be salvaged.