Cleanup dead code re: upgrade in http_req_sock.
Generalize the upgrade state and rename it to paused state. After
reflection, there is no need to store tuples in the queue because if we
are in the paused state, we know the upgrade should be the last
response.

Fixes a bug in reset handler where queued responses were not deleted.