Make a stronger attempt at request/response replaying.
Replay responses to a fake URL that appears to be from the original
Host, from the browser's point of view. In this way, relative URLs will
point to the original host. We must first redirect to a randomly
generated page name on the host, which we also must remember to hijack
later.