Bugfix/cleanup for http_res_sock.
I remembered I changed the structure of the Target tuple to be:
  {Host,Port,Secure}
where Secure is true if the connection should use SSL/TLS and false if
it is unencrypted. The problem is I forgot to update request_target
and http_res_sock to use this new tuple.

I also deleted some unused code that I just added and very quickly made
redundant.