Fixed certificate forgery after many hours.
Elliptic curve crypto stopped working. The curve I was using was no
longer supported, or considered dangerously insecure perhaps. I switched
to using RSA instead. On the way I found that the parameters for RSA are
more strict (you need a NULL for the signature parameters). This was
causing a BAD_DER error to be emitted by Firefox.

- Refactored forger_lib quite a bit while switching to RSA.
- Decided to use a simple func (connect_tunnel) instead of sending a
  cast message inside of http_req_sock.