Create flutter app under ui/
I'm not sure what to do with the other platforms but I might be able to
get more fancy later on. For now the web application is only available
when it is served from *inside* of piroxy. The proxy is hijacking the
connection when it see a request to "piroxy". Later on I might be able
to have a separate web server node that is connected to the proxy node.
The apps would then communicate in mostly the same way.